adding local search code
