+++ prtpy/packing/triplet_algo/triplet_planner.py
@@ -9,526 +9,11 @@ import logging
 from .models import SolverData, TripletSearchContext
 from .triplet_backtracker import TripletBacktracker
 from .base import USE_IMPROVEMENT_HEURISTIC
-# from .triplet_refiner import TripletRefiner
+from .triplet_local_search import TripletLocalSearch
 
 logger = logging.getLogger("trialgo.triplet_algo")
 
 WeightType = int
-
-
-
-
-
-
-
-
-
-# from __future__ import annotations
-from dataclasses import dataclass, field
-from typing import Tuple, List, Set, Dict, Optional, Any
-import heapq
-# from .triplet_planner import TripletPlanner
-# import logging
-
-# logger = logging.getLogger("trialgo.heuristic")
-
-
-# @dataclass
-# class Stats:
-#     passes: int = 0
-#     distance: int = 0
-#     node_count: int = 0
-#     saved_count: int = 0
-#     skip1_count: int = 0
-#     skip2_count: int = 0
-
-
-# @dataclass
-# class TripletInfo:
-#     triplet: Tuple[int, int, int]
-#     used_count: int = 0
-
-
-# @dataclass
-# class GroupInfo:
-#     left: int = 0
-#     triplets: Set[int] = field(default_factory=set)
-
-
-# class TripletRefiner:
-#     def __init__(
-#         self, orig_chosen_triplet_indices: List[int], level_a: 'TripletPlanner'
-#     ) -> None:
-#         self.orig_chosen_triplet_indices = list(orig_chosen_triplet_indices)
-#         # initialize triplets
-#         T = len(level_a.triplets_abc)
-#         self.triplets: List[TripletInfo] = [
-#             TripletInfo(level_a.triplets_abc[t], 0) for t in range(T)
-#         ]
-
-#         # initialize groups
-#         G = len(level_a.groups)
-#         self.groups: List[GroupInfo] = [
-#             GroupInfo(level_a.group_cardinality[g]) for g in range(G)
-#         ]
-#         # assign triplet indices to groups
-#         # for t, info in enumerate(self.triplets):
-#         for t in range(T):
-#             # a,b,c = info.triplet
-#             a, b, c = self.triplets[t].triplet
-#             self.groups[a].triplets.add(t)
-#             self.groups[b].triplets.add(t)
-#             self.groups[c].triplets.add(t)
-
-#         # apply original chosen
-#         for t in self.orig_chosen_triplet_indices:
-#             self.triplets[t].used_count += 1
-#             a, b, c = self.triplets[t].triplet
-#             self.groups[a].left -= 1
-#             self.groups[b].left -= 1
-#             self.groups[c].left -= 1
-
-#         self.stats = Stats()
-
-#     class Node:
-#         __slots__ = (
-#             "parent",
-#             "prev",
-#             "triplet_index",
-#             "is_add",
-#             "infeas_count",
-#             "extra_items",
-#             "group_left",
-#             "triplet_used_count",
-#         )
-
-#         def __init__(
-#             self,
-#             parent: 'TripletRefiner',
-#             prev: Optional['TripletRefiner.Node'] = None,
-#             is_add: bool = False,
-#             triplet_index: int = 0,
-#         ) -> None:
-#             self.parent = parent
-#             self.prev = prev
-#             self.triplet_index = triplet_index
-#             self.is_add = is_add
-#             # inherit mutable state or start fresh
-#             if prev is None:
-#                 self.infeas_count = 0
-#                 self.extra_items = 0
-#                 self.group_left: Dict[int, int] = {}
-#                 self.triplet_used_count: Dict[int, int] = {}
-#             else:
-#                 self.infeas_count = prev.infeas_count
-#                 self.extra_items = prev.extra_items
-#                 self.group_left = dict(prev.group_left)
-#                 self.triplet_used_count = dict(prev.triplet_used_count)
-#                 a, b, c = parent.triplets[triplet_index].triplet
-#                 if is_add:
-#                     self.addTriplet(triplet_index)
-#                     self._adjust_group(a, -1)
-#                     self._adjust_group(b, -1)
-#                     self._adjust_group(c, -1)
-#                 else:
-#                     self.removeTriplet(triplet_index)
-#                     self._adjust_group(a, +1)
-#                     self._adjust_group(b, +1)
-#                     self._adjust_group(c, +1)
-
-#         def __lt__(self, other: 'TripletRefiner.Node') -> bool:
-#             # fewer infeas, then more extra_items, then group_left dict comparison
-#             if self.infeas_count != other.infeas_count:
-#                 return self.infeas_count < other.infeas_count
-#             if self.extra_items != other.extra_items:
-#                 return self.extra_items > other.extra_items
-#             return sorted(self.group_left.items()) < sorted(other.group_left.items())
-
-#         def _adjust_group(self, g: int, delta: int) -> None:
-#             base = self.parent.groups[g].left
-#             left = self.group_left.get(g, base)
-#             new = left + delta
-#             if delta < 0 and new < 0:
-#                 self.infeas_count += 1
-#             if delta > 0 and new <= 0:
-#                 self.infeas_count -= 1
-#             if new == base:
-#                 self.group_left.pop(g, None)
-#             else:
-#                 self.group_left[g] = new
-#             self.extra_items += -delta
-
-#         def addTriplet(self, t: int) -> None:
-#             base = self.parent.triplets[t].used_count
-#             count = self.triplet_used_count.get(t, base)
-#             self.triplet_used_count[t] = count + 1
-
-#         def removeTriplet(self, t: int) -> None:
-#             base = self.parent.triplets[t].used_count
-#             count = self.triplet_used_count.get(t, base)
-#             self.triplet_used_count[t] = count - 1
-
-#         def getTripletUsedCount(self, t: int) -> int:
-#             return self.triplet_used_count.get(t, self.parent.triplets[t].used_count)
-
-#         def getGroupLeft(self, g: int) -> int:
-#             return self.group_left.get(g, self.parent.groups[g].left)
-
-#         def applySolutionOnParent(self) -> None:
-#             # collect path
-#             path: List[TripletRefiner.Node] = []
-#             node = self
-#             while node.prev:
-#                 path.append(node)
-#                 node = node.prev
-
-#             for n in reversed(path):
-#                 t = n.triplet_index
-#                 if n.is_add:
-#                     self.parent.triplets[t].used_count += 1
-#                     a, b, c = self.parent.triplets[t].triplet
-#                     self.parent.groups[a].left -= 1
-#                     self.parent.groups[b].left -= 1
-#                     self.parent.groups[c].left -= 1
-#                 else:
-#                     self.parent.triplets[t].used_count -= 1
-#                     a, b, c = self.parent.triplets[t].triplet
-#                     self.parent.groups[a].left += 1
-#                     self.parent.groups[b].left += 1
-#                     self.parent.groups[c].left += 1
-#             self.parent.stats.distance += len(path)
-
-#     def perform(self, max_infeas_count: int, max_steps: int) -> Tuple[bool, int]:
-#         self.stats.passes += 1
-
-#         T = len(self.triplets)
-#         G = len(self.groups)
-
-#         # priority queue and membership
-#         nodes_remaining: List[TripletRefiner.Node] = []
-#         nodes_all_keys: Set[Any] = set()
-
-#         # helper to fingerprint a node
-#         def key(n: TripletRefiner.Node):
-#             return (
-#                 n.infeas_count,
-#                 -n.extra_items,
-#                 frozenset(n.group_left.items()),
-#                 frozenset(n.triplet_used_count.items()),
-#             )
-
-#         # root
-#         root = TripletRefiner.Node(self)
-#         heapq.heappush(nodes_remaining, root)
-#         nodes_all_keys.add(key(root))
-#         nodes_all_count = 1
-
-#         for step_count in range(max_steps):
-#             if not nodes_remaining:
-#                 self.stats.node_count = step_count + 1
-#                 self.stats.saved_count = nodes_all_count
-#                 return False, step_count
-
-#             node = heapq.heappop(nodes_remaining)
-
-#             if node.infeas_count == 0 and node.extra_items > 0:
-#                 node.applySolutionOnParent()
-#                 self.stats.node_count = step_count + 1
-#                 self.stats.saved_count = nodes_all_count
-#                 return True, step_count
-
-#             # build next triplets
-#             next_triplets: Set[int] = set()
-#             is_add = node.infeas_count == 0
-#             if is_add:
-#                 for g in range(G):
-#                     if node.getGroupLeft(g) > 0:
-#                         next_triplets |= self.groups[g].triplets
-#             else:
-#                 for g, left in node.group_left.items():
-#                     if left < 0:
-#                         next_triplets |= self.groups[g].triplets
-
-#             # expand
-#             for t in next_triplets:
-#                 if not is_add and node.getTripletUsedCount(t) == 0:
-#                     continue
-#                 nxt = TripletRefiner.Node(self, node, is_add, t)
-#                 if nxt.infeas_count > max_infeas_count:
-#                     self.stats.skip1_count += 1
-#                     continue
-#                 k = key(nxt)
-#                 if k in nodes_all_keys:
-#                     self.stats.skip2_count += 1
-#                     continue
-#                 heapq.heappush(nodes_remaining, nxt)
-#                 nodes_all_keys.add(k)
-#                 nodes_all_count += 1
-
-#         self.stats.node_count += max_steps
-#         self.stats.saved_count += nodes_all_count
-#         return False, max_steps
-
-#     def getFinalTripletCount(self) -> int:
-#         return sum(info.used_count for info in self.triplets)
-
-#     def getFinalTripletIndices(self) -> List[int]:
-#         return [
-#             t for t, info in enumerate(self.triplets) for _ in range(info.used_count)
-#         ]
-
-#     def getFinalTriplets(self) -> List[Tuple[int, int, int]]:
-#         return [self.triplets[t].triplet for t in self.getFinalTripletIndices()]
-
-#     def getStats(self) -> Stats:
-#         return self.stats
-
-
-
-
-from typing import List, Tuple, Set, Dict, Optional
-import heapq
-import itertools
-
-
-class LevelC:
-    class Stats:
-        def __init__(self):
-            self.passes = 0
-            self.distance = 0
-            self.node_count = 0
-            self.saved_count = 0
-            self.skip1_count = 0
-            self.skip2_count = 0
-
-    class TripletInfo:
-        def __init__(self, triplet: Tuple[int, int, int]):
-            self.triplet = triplet
-            self.used_count = 0
-
-    class GroupInfo:
-        def __init__(self, left: int):
-            self.left = left
-            self.triplets: Set[int] = set()
-
-    class Node:
-        def __init__(self, parent: 'LevelC', prev: Optional['LevelC.Node'] = None,
-                     is_add: Optional[bool] = None, triplet_index: Optional[int] = None):
-            self.parent = parent
-            self.prev = prev
-            self.triplet_index = triplet_index
-            self.is_add = is_add
-            self.infeas_count = prev.infeas_count if prev else 0
-            self.extra_items = prev.extra_items if prev else 0
-            self.group_left = dict(prev.group_left) if prev else {}
-            self.triplet_used_count = dict(prev.triplet_used_count) if prev else {}
-
-            if triplet_index is not None:
-                a, b, c = parent.triplets[triplet_index].triplet
-                if is_add:
-                    self.addTriplet(triplet_index)
-                    self.add(a)
-                    self.add(b)
-                    self.add(c)
-                else:
-                    self.removeTriplet(triplet_index)
-                    self.remove(a)
-                    self.remove(b)
-                    self.remove(c)
-
-        def getTripletUsedCount(self, t: int) -> int:
-            return self.triplet_used_count.get(t, self.parent.triplets[t].used_count)
-
-        def getGroupLeft(self, g: int) -> int:
-            return self.group_left.get(g, self.parent.groups[g].left)
-
-        def addTriplet(self, t: int):
-            self.triplet_used_count[t] = self.getTripletUsedCount(t) + 1
-
-        def removeTriplet(self, t: int):
-            self.triplet_used_count[t] = self.getTripletUsedCount(t) - 1
-
-        def add(self, g: int):
-            current = self.group_left.get(g, self.parent.groups[g].left)
-            new_left = current - 1
-
-            if new_left < 0:
-                self.infeas_count += 1
-
-            if new_left == self.parent.groups[g].left:
-                self.group_left.pop(g, None)
-            else:
-                self.group_left[g] = new_left
-
-            self.extra_items += 1
-
-        def remove(self, g: int):
-            current = self.group_left.get(g, self.parent.groups[g].left)
-            new_left = current + 1
-
-            if new_left <= 0:
-                self.infeas_count -= 1
-
-            if new_left == self.parent.groups[g].left:
-                self.group_left.pop(g, None)
-            else:
-                self.group_left[g] = new_left
-
-            self.extra_items -= 1
-
-        def __lt__(self, other: 'LevelC.Node'):
-            if self.infeas_count != other.infeas_count:
-                return self.infeas_count < other.infeas_count
-            if self.extra_items != other.extra_items:
-                return self.extra_items > other.extra_items
-            return sorted(self.group_left.items()) < sorted(other.group_left.items())
-
-        def __eq__(self, other):
-            return (self.infeas_count == other.infeas_count and
-                    self.extra_items == other.extra_items and
-                    self.group_left == other.group_left and
-                    self.triplet_used_count == other.triplet_used_count)
-
-        def __hash__(self):
-            return hash((
-                self.infeas_count,
-                self.extra_items,
-                frozenset(self.group_left.items()),
-                frozenset(self.triplet_used_count.items())
-            ))
-
-        def applySolutionOnParent(self):
-            nodes = []
-            node = self
-            while node.prev is not None:
-                nodes.append(node)
-                node = node.prev
-            for node in reversed(nodes):
-                a, b, c = node.parent.triplets[node.triplet_index].triplet
-                if node.is_add:
-                    node.parent.triplets[node.triplet_index].used_count += 1
-                    node.parent.groups[a].left -= 1
-                    node.parent.groups[b].left -= 1
-                    node.parent.groups[c].left -= 1
-                else:
-                    node.parent.triplets[node.triplet_index].used_count -= 1
-                    node.parent.groups[a].left += 1
-                    node.parent.groups[b].left += 1
-                    node.parent.groups[c].left += 1
-            self.parent.stats.distance += len(nodes)
-
-    def __init__(self, orig_chosen_triplet_indices: List[int], level_a):
-        print("int levelc init")
-        self.orig_chosen_triplet_indices = orig_chosen_triplet_indices
-        self.triplets: List[LevelC.TripletInfo] = []
-        self.groups: List[LevelC.GroupInfo] = []
-        self.stats = LevelC.Stats()
-
-        T = len(level_a.triplets_abc)
-        self.triplets = [LevelC.TripletInfo(t) for t in level_a.triplets_abc]
-
-        G = len(level_a.groups)
-        self.groups = [LevelC.GroupInfo(card) for card in level_a.group_cardinality]
-
-        for t in range(T):
-            a, b, c = self.triplets[t].triplet
-            self.groups[a].triplets.add(t)
-            self.groups[b].triplets.add(t)
-            self.groups[c].triplets.add(t)
-
-        for t in orig_chosen_triplet_indices:
-            self.triplets[t].used_count += 1
-            a, b, c = self.triplets[t].triplet
-            self.groups[a].left -= 1
-            self.groups[b].left -= 1
-            self.groups[c].left -= 1
-
-    def perform(self, max_infeas_count: int, max_steps: int) -> Tuple[bool, int]:
-        self.stats.passes += 1
-        T = len(self.triplets)
-        G = len(self.groups)
-
-        counter = itertools.count()
-        nodes_all = set()
-        nodes_remaining = []
-
-        root = LevelC.Node(self)
-        heapq.heappush(nodes_remaining, (0, next(counter), root))
-        nodes_all.add(root)
-
-        for step_count in range(max_steps):
-            if not nodes_remaining:
-                self.stats.node_count += step_count + 1
-                self.stats.saved_count += len(nodes_all)
-                return False, step_count
-
-            _, _, node = heapq.heappop(nodes_remaining)
-
-            if node.infeas_count == 0 and node.extra_items > 0:
-                node.applySolutionOnParent()
-                self.stats.node_count += step_count + 1
-                self.stats.saved_count += len(nodes_all)
-                return True, step_count
-
-            next_triplet_indices = set()
-            is_add = node.infeas_count == 0
-            if is_add:
-                for g in range(G):
-                    if node.getGroupLeft(g) > 0:
-                        next_triplet_indices.update(self.groups[g].triplets)
-            else:
-                for g, left in node.group_left.items():
-                    if left < 0:
-                        next_triplet_indices.update(self.groups[g].triplets)
-
-            for t in next_triplet_indices:
-                if not is_add and node.getTripletUsedCount(t) == 0:
-                    continue
-                next_node = LevelC.Node(self, node, is_add, t)
-                if next_node.infeas_count > max_infeas_count:
-                    self.stats.skip1_count += 1
-                    continue
-                if next_node in nodes_all:
-                    self.stats.skip2_count += 1
-                    continue
-                heapq.heappush(nodes_remaining, (next_node.infeas_count, next(counter), next_node))
-                nodes_all.add(next_node)
-
-        self.stats.node_count += max_steps
-        self.stats.saved_count += len(nodes_all)
-        return False, max_steps
-
-    def getFinalTripletCount(self) -> int:
-        return sum(t.used_count for t in self.triplets)
-
-    def getFinalTripletIndices(self) -> List[int]:
-        final_triplet_indices = []
-        for t in range(len(self.triplets)):
-            final_triplet_indices.extend([t] * self.triplets[t].used_count)
-        return final_triplet_indices
-
-    def getFinalTriplets(self) -> List[Tuple[int, int, int]]:
-        return [self.triplets[t].triplet for t in self.getFinalTripletIndices()]
-
-    def getStats(self) -> 'LevelC.Stats':
-        return self.stats
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 class TripletPlanner:
@@ -848,14 +333,14 @@ class TripletPlanner:
             self.answer.a_cases_investigated = case_index + 1
             self.answer.winning_branches = stats.winning_branches
 
-            level_improvement = LevelC(level_backtrack.get_chosen_triplet_indices(), self)
+            level_improvement = TripletLocalSearch(level_backtrack.get_chosen_triplet_indices(), self.triplets_abc, self.group_cardinality)
             improvement_final_success = False
 
             while True:
-                if level_improvement.getFinalTripletCount() == A:
+                if level_improvement.get_final_triplet_count() == A:
                     logger.debug("Improvement successful, full solution found")
                     self.answer.success = True
-                    self.algorithm_chosen_triplets = level_improvement.getFinalTriplets()
+                    self.algorithm_chosen_triplets = level_improvement.get_final_triplets()
                     improvement_final_success = True
                     break
 
@@ -865,7 +350,7 @@ class TripletPlanner:
                     logger.debug("Improvement failed to proceed further")
                     break
 
-            stats = level_improvement.getStats()
+            stats = level_improvement.get_stats()
             logger.debug(f"Improvement stats: {stats}")
 
             self.answer.improvement_passes = stats.passes
